requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.13.3
azure-storage-blob==12.16.0
azure-data-tables==12.4.3
pika==1.3.1
azure-eventhub==5.11.2
aiohttp==3.8.4
azure-cosmosdb-table